refactor(msa): drop test-only q-major + plain-CSR paths; rename to canonical names
The q-major sparse-attention kernel and the plain tuple-returning CSR builder
were test-only scaffolding, never on the M3 serving path. Remove them and
promote the kv-major / schedule variants to the canonical public names:

  - msa_sparse_attention_kvmajor      -> msa_sparse_attention
  - SparseAttentionForwardKvMajorSm12x -> SparseAttentionForwardSm12x
    (file sparse_fwd_kvmajor_sm12x.py  -> sparse_fwd_sm12x.py)
  - msa_build_k2q_csr_schedule        -> msa_build_k2q_csr  (returns MsaAttentionSchedule)
  - benchmark MSASparseAttentionKvMajor -> MSASparseAttention
  - trace template msa_sparse_attention_kvmajor_trace -> msa_sparse_attention_trace

Also scrub the now-stale naming the removal left behind: docstrings/comments
that named the deleted q-major sibling or the build_k2q_csr_schedule symbol,
the dead _MAX_KV_BLOCKS SMEM-selection-map constant (only the q-major path
used it), the "kvmajor"/"(schedule variant)" wording in kernel/benchmark
docs, the cache-key literal, and the "kvmajor" test labels.

No kernel math changes. Verified: pre-commit clean, tests/msa_ops/ green,
tests/trace/ 694 passed.
