+++ benchmarks/routines/flashinfer_benchmark_utils.py
@@ -289,7 +289,7 @@ benchmark_apis = {
     ],
     "sparse_attention": [
         "MSAProxyScore",
-        "MSASparseAttentionKvMajor",
+        "MSASparseAttention",
         "MSASparseDecode",
         "MSAPipeline",
     ],
