+++ flashinfer/msa_ops/cute_dsl/__init__.py
@@ -1,9 +1,7 @@
 from .sparse_combine_sm12x import SparseCombineSm12x
-from .sparse_fwd_kvmajor_sm12x import SparseAttentionForwardKvMajorSm12x
 from .sparse_fwd_sm12x import SparseAttentionForwardSm12x
 
 __all__ = [
-    "SparseAttentionForwardKvMajorSm12x",
     "SparseAttentionForwardSm12x",
     "SparseCombineSm12x",
 ]
